requirements.txt
pyyaml==6.0
dirtyjson==1.0.8